requirements.txt
django>=4.2
celery
redis
google-auth
google-auth-oauthlib
google-api-python-client
djangorestframework
python-decouple